+++ smacc2_performance_tools/performance_tests/sm_coretest_transition_speed_1/include/sm_coretest_transition_speed_1/sm_coretest_transition_speed_1.hpp
@@ -30,11 +30,11 @@ struct State1;
 struct State2;
 
 //VARIABLES - shared between states (using "_<name>_"-syntax to make this obvious)
-static unsigned int _counter_ = 1;
+static uint64_t _counter_ = 1;
 std::shared_ptr<rclcpp::Node> _node_;
 rclcpp::Time _start_time_;
 
-unsigned int _sum_of_iterations_ = 0.0;
+uint64_t _sum_of_iterations_ = 0.0;
 double _sum_of_elapsed_time_ = 0.0;
 
 //--------------------------------------------------------------------
