+++ smacc2_sm_reference_library/sm_husky_barrel_search_1/src/sm_husky_barrel_search_1/gz_plugins/FlashLightPlugin.cc
@@ -511,7 +511,7 @@ void FlashLightSetting::SetColor(const ignition::math::Color &_color)
 }
 
 //////////////////////////////////////////////////
-unsigned int FlashLightSetting::BlockCount()
+uint64_t FlashLightSetting::BlockCount()
 {
   return this->dataPtr->blocks.size();
 }
